refactor(account-chests): Split up account actions and move townhall query down
