+++ apps/website/src/queries/townhall.ts
@@ -1,15 +1,11 @@
 import { db } from "@/db";
 
 export const getTownhallByLevel = async (level: number) => {
-  console.log("Fetching townhall with level:", level);
-
   const townhall = await db
     .selectFrom("townhall")
     .select(["townhall.id", "townhall.level"])
     .where("townhall.level", "=", level)
     .executeTakeFirst();
-
-  console.log("Queried townhall:", townhall);
 
   return townhall;
 };
